connected NameInput component to Form
