+++ code/src/components/Form.js
@@ -1,29 +1,12 @@
-import React, { useState } from "react";
-
-
-
+import React from "react";
+import NameInput from "./NameInput";
 
 
 const Form = () => {
 
-
-    const [name, setName] = useState('')
-
-    const onNameChange = (event) => {
-        console.log(event.target.value)
-        setName(event.target.value);
-    }
     return (
         <form>
-            <div>
-                <label htmlFor="name">What's your name?</label>
-                <input
-                    id="name"
-                    type="text"
-                    value={name}
-                    onChange={onNameChange}
-                />
-            </div>
+            <NameInput />
         </form>
     )
 }
